refactor(review): extract core review service
