+++ src/commands/review/mod.rs
@@ -12,35 +12,30 @@
 //!
 //! See: https://github.com/Extra-Chill/homeboy/issues/1500
 
-use chrono::Utc;
 use clap::Args;
 use serde::Serialize;
 use serde_json::Value;
 use std::path::Path;
-use std::process::Command;
 
-use homeboy::core::ci_profile::{self, CiRunOutput, CiRunSelection};
+use homeboy::core::ci_profile::{self, CiRunSelection};
 use homeboy::core::code_audit::AuditCommandOutput;
 use homeboy::core::engine::execution_context::{self, ResolveOptions};
-use homeboy::core::execution::{self, PlanExecutionRun};
 use homeboy::core::extension::lint::LintCommandOutput;
 use homeboy::core::extension::test::TestCommandOutput;
-use homeboy::core::finding::HomeboyFinding;
 use homeboy::core::git;
-use homeboy::core::plan::{HomeboyPlan, PlanStep};
+use homeboy::core::plan::PlanStep;
 use homeboy::core::quality::{build_quality_plan, QualityPlanOptions};
-use homeboy::core::ObservationOutputMetadata;
+use homeboy::core::review::{
+    self, ReviewArtifactFindings, ReviewCommandOutput, ReviewOutputInput, ReviewService,
+    ReviewStage, ReviewStages,
+};
 
 use super::parse_key_val;
 use super::utils::args::{BaselineArgs, ExtensionOverrideArgs, PositionalComponentArgs};
 use super::{audit, lint, test, CmdResult, GlobalArgs};
 
-mod artifact_findings;
 mod observation;
 pub(super) mod raw_output;
-mod render;
-
-use artifact_findings::ReviewArtifactFindings;
 
 #[derive(Args, Debug, Clone)]
 pub struct ReviewArgs {
@@ -92,94 +87,6 @@ pub fn is_markdown_mode(args: &ReviewArgs) -> bool {
     args.report.as_deref() == Some("pr-comment")
 }
 
-/// Per-stage section of the consolidated review output.
-#[derive(Serialize)]
-pub struct ReviewStage<T: Serialize> {
-    /// Stage name (`"audit"`, `"lint"`, `"test"`).
-    pub stage: String,
-    /// Whether the stage ran or was skipped.
-    pub ran: bool,
-    /// Stage-level pass/fail (only meaningful when `ran` is true).
-    pub passed: bool,
-    /// Stage exit code (0 when skipped).
-    pub exit_code: i32,
-    /// Number of findings the stage reported (audit findings, lint findings,
-    /// test failures). Always 0 when skipped or stage-internal counts unavailable.
-    pub finding_count: usize,
-    /// Human-readable hint pointing to the per-stage command for deep dive.
-    pub hint: String,
-    /// Skip reason (only present when `ran` is false).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub skipped_reason: Option<String>,
-    /// Full structured output from the underlying command. None if skipped.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub output: Option<T>,
-}
-
-/// Top-level summary block — what a reviewer would skim first.
-#[derive(Serialize)]
-pub struct ReviewSummary {
-    /// True when every stage that ran exited 0.
-    pub passed: bool,
-    /// Top-line status string.
-    pub status: String,
-    /// Component label.
-    pub component: String,
-    /// Scope mode applied: `"changed-since"`, `"changed-only"`, or `"full"`.
-    pub scope: String,
-    /// The git ref passed to `--changed-since`, if any.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub changed_since: Option<String>,
-    /// Total findings across all stages that ran.
-    pub total_findings: usize,
-    /// Count of files in the changed set (None when not in scoped mode).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub changed_file_count: Option<usize>,
-    /// Top-level hints (e.g., empty changeset, scope warnings).
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub hints: Vec<String>,
-}
-
-/// Unified output envelope for the review command.
-#[derive(Serialize)]
-pub struct ReviewCommandOutput {
-    pub command: String,
-    pub plan: HomeboyPlan,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub observation: Option<ObservationOutputMetadata>,
-    pub artifact: ReviewArtifact,
-    pub summary: ReviewSummary,
-    pub audit: ReviewStage<AuditCommandOutput>,
-    pub lint: ReviewStage<LintCommandOutput>,
-    pub test: ReviewStage<TestCommandOutput>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ci_profile: Option<ReviewStage<CiRunOutput>>,
-}
-
-/// Stable machine-readable artifact for automated PR review consumers.
-#[derive(Serialize, Clone)]
-pub struct ReviewArtifact {
-    pub schema: String,
-    pub component: String,
-    pub status: String,
-    pub generated_at: String,
-    pub base_ref: String,
-    pub head_ref: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub observation: Option<ObservationOutputMetadata>,
-    pub commands: Vec<ReviewArtifactCommand>,
-}
-
-#[derive(Serialize, Clone)]
-pub struct ReviewArtifactCommand {
-    pub name: String,
-    pub status: String,
-    pub exit_code: i32,
-    pub summary: String,
-    pub findings: Vec<HomeboyFinding>,
-    pub artifacts: Vec<Value>,
-}
-
 struct ReviewStageDescriptor<Args, Output: Serialize + ReviewArtifactFindings> {
     name: &'static str,
     include_changed_only_scope: bool,
@@ -189,9 +96,9 @@ struct ReviewStageDescriptor<Args, Output: Serialize + ReviewArtifactFindings> {
 }
 
 enum ReviewStageRun {
-    Audit(ReviewStage<AuditCommandOutput>, i32),
-    Lint(ReviewStage<LintCommandOutput>, i32),
-    Test(ReviewStage<TestCommandOutput>, i32),
+    Audit(ReviewStage<AuditCommandOutput>),
+    Lint(ReviewStage<LintCommandOutput>),
+    Test(ReviewStage<TestCommandOutput>),
 }
 
 impl<Args, Output: Serialize + ReviewArtifactFindings> ReviewStageDescriptor<Args, Output> {
@@ -225,16 +132,6 @@ impl<Args, Output: Serialize + ReviewArtifactFindings> ReviewStageDescriptor<Arg
     }
 }
 
-fn execute_review_plan_steps<R, Dispatch>(
-    steps: &[PlanStep],
-    dispatch: Dispatch,
-) -> homeboy::core::Result<PlanExecutionRun<R>>
-where
-    Dispatch: FnMut(&PlanStep) -> homeboy::core::Result<Option<R>>,
-{
-    execution::execute_plan_steps(steps, dispatch, |_| false)
-}
-
 fn dispatch_review_plan_step(
     step: &PlanStep,
     args: &ReviewArgs,
@@ -250,8 +147,8 @@ fn dispatch_review_plan_step(
                 run: audit::run,
                 finding_count: audit_finding_count,
             };
-            let (stage, exit_code) = descriptor.execute(args, global, component_label)?;
-            Ok(Some(ReviewStageRun::Audit(stage, exit_code)))
+            let (stage, _) = descriptor.execute(args, global, component_label)?;
+            Ok(Some(ReviewStageRun::Audit(stage)))
         }
         "review.lint" => {
             let descriptor = ReviewStageDescriptor {
@@ -261,8 +158,8 @@ fn dispatch_review_plan_step(
                 run: lint::run,
                 finding_count: lint_finding_count,
             };
-            let (stage, exit_code) = descriptor.execute(args, global, component_label)?;
-            Ok(Some(ReviewStageRun::Lint(stage, exit_code)))
+            let (stage, _) = descriptor.execute(args, global, component_label)?;
+            Ok(Some(ReviewStageRun::Lint(stage)))
         }
         "review.test" => {
             let descriptor = ReviewStageDescriptor {
@@ -272,8 +169,8 @@ fn dispatch_review_plan_step(
                 run: test::run,
                 finding_count: test_finding_count,
             };
-            let (stage, exit_code) = descriptor.execute(args, global, component_label)?;
-            Ok(Some(ReviewStageRun::Test(stage, exit_code)))
+            let (stage, _) = descriptor.execute(args, global, component_label)?;
+            Ok(Some(ReviewStageRun::Test(stage)))
         }
         other => Err(homeboy::core::Error::internal_unexpected(format!(
             "review quality plan contains unsupported executable step '{other}'"
@@ -328,56 +225,23 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
         });
         let observation_metadata = review_observation.as_ref().map(|o| o.output_metadata());
 
-        let mut artifact = ReviewArtifact {
-            schema: "homeboy/review/v1".to_string(),
-            component: component_label.clone(),
-            status: "skipped".to_string(),
-            generated_at: generated_at_now(),
-            base_ref: args.changed_since.clone().unwrap_or_default(),
-            head_ref: git_ref(&source_path, "HEAD").unwrap_or_default(),
-            observation: None,
-            commands: vec![
-                artifact_command(&stage_skipped::<Value>("audit", "no files changed")),
-                artifact_command(&stage_skipped::<Value>("lint", "no files changed")),
-                artifact_command(&stage_skipped::<Value>("test", "no files changed")),
-            ],
-        };
-        if args.ci_profile.is_some() {
-            artifact
-                .commands
-                .push(artifact_command(&stage_skipped::<Value>(
-                    "ci",
-                    "no files changed",
-                )));
-        }
-        artifact.observation = observation_metadata.clone();
-
-        let output = ReviewCommandOutput {
-            command: "review".to_string(),
-            plan: build_quality_plan(QualityPlanOptions::skipped_review(
-                &component_label,
-                "no files changed",
-            )),
-            observation: observation_metadata,
-            artifact,
-            summary: ReviewSummary {
-                passed: true,
-                status: "passed".to_string(),
+        let output = ReviewService::skipped_output(
+            ReviewOutputInput {
                 component: component_label.clone(),
+                plan: build_quality_plan(QualityPlanOptions::skipped_review(
+                    &component_label,
+                    "no files changed",
+                )),
+                observation: observation_metadata,
                 scope: scope.clone(),
                 changed_since: args.changed_since.clone(),
-                total_findings: 0,
                 changed_file_count: Some(0),
+                head_ref: review::git_ref(&source_path, "HEAD").unwrap_or_default(),
                 hints: vec![message],
             },
-            audit: stage_skipped("audit", "no files changed"),
-            lint: stage_skipped("lint", "no files changed"),
-            test: stage_skipped("test", "no files changed"),
-            ci_profile: args
-                .ci_profile
-                .as_ref()
-                .map(|_| stage_skipped("ci", "no files changed")),
-        };
+            "no files changed",
+            args.ci_profile.is_some(),
+        );
         observation::finish_success(review_observation, &output, 0);
         return Ok((output, 0));
     }
@@ -394,13 +258,10 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
     let mut top_hints: Vec<String> = Vec::new();
 
     let mut audit_stage = None;
-    let mut audit_exit = 0;
     let mut lint_stage = None;
-    let mut lint_exit = 0;
     let mut test_stage = None;
-    let mut test_exit = 0;
 
-    let stage_run = match execute_review_plan_steps(&quality_plan.steps, |step| {
+    let stage_run = match review::execute_review_plan_steps(&quality_plan.steps, |step| {
         dispatch_review_plan_step(step, &args, global, &component_label)
     }) {
         Ok(run) => run,
@@ -412,17 +273,14 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
 
     for stage_run in stage_run.results {
         match stage_run {
-            ReviewStageRun::Audit(stage, exit_code) => {
+            ReviewStageRun::Audit(stage) => {
                 audit_stage = Some(stage);
-                audit_exit = exit_code;
             }
-            ReviewStageRun::Lint(stage, exit_code) => {
+            ReviewStageRun::Lint(stage) => {
                 lint_stage = Some(stage);
-                lint_exit = exit_code;
             }
-            ReviewStageRun::Test(stage, exit_code) => {
+            ReviewStageRun::Test(stage) => {
                 test_stage = Some(stage);
-                test_exit = exit_code;
             }
         }
     }
@@ -430,7 +288,7 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
     let audit_stage = audit_stage.expect("review quality plan must include audit stage");
     let lint_stage = lint_stage.expect("review quality plan must include lint stage");
     let test_stage = test_stage.expect("review quality plan must include test stage");
-    let (ci_profile_stage, ci_profile_exit) = match args.ci_profile.as_ref() {
+    let ci_profile_stage = match args.ci_profile.as_ref() {
         Some(profile) => {
             let ctx = execution_context::resolve(&ResolveOptions {
                 component_id: args.comp.component.clone(),
@@ -471,36 +329,10 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
                 skipped_reason: None,
                 output: Some(output),
             };
-            (Some(stage), exit_code)
+            Some(stage)
         }
-        None => (None, 0),
+        None => None,
     };
-
-    // Aggregate
-    let overall_passed = audit_stage.passed
-        && lint_stage.passed
-        && test_stage.passed
-        && ci_profile_stage
-            .as_ref()
-            .map(|stage| stage.passed)
-            .unwrap_or(true);
-    let overall_exit = if overall_passed {
-        0
-    } else if [audit_exit, lint_exit, test_exit, ci_profile_exit]
-        .iter()
-        .any(|&c| c >= 2)
-    {
-        2
-    } else {
-        1
-    };
-    let total_findings = audit_stage.finding_count
-        + lint_stage.finding_count
-        + test_stage.finding_count
-        + ci_profile_stage
-            .as_ref()
-            .map(|stage| stage.finding_count)
-            .unwrap_or(0);
 
     if args.changed_only {
         top_hints.push(
@@ -508,46 +340,25 @@ pub fn run(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<ReviewCommandOutp
         );
     }
 
-    let summary = ReviewSummary {
-        passed: overall_passed,
-        status: if overall_passed { "passed" } else { "failed" }.to_string(),
-        component: component_label.clone(),
-        scope,
-        changed_since: args.changed_since.clone(),
-        total_findings,
-        changed_file_count,
-        hints: top_hints,
-    };
-
-    let mut commands = vec![
-        artifact_command(&audit_stage),
-        artifact_command(&lint_stage),
-        artifact_command(&test_stage),
-    ];
-    if let Some(ref stage) = ci_profile_stage {
-        commands.push(artifact_command(stage));
-    }
-
-    let mut artifact = build_artifact(
-        &component_label,
-        args.changed_since.as_deref().unwrap_or(""),
-        git_ref(&source_path, "HEAD").unwrap_or_default().as_str(),
-        commands,
-    );
     let observation_metadata = review_observation.as_ref().map(|o| o.output_metadata());
-    artifact.observation = observation_metadata.clone();
-
-    let output = ReviewCommandOutput {
-        command: "review".to_string(),
-        plan: quality_plan,
-        observation: observation_metadata,
-        artifact,
-        summary,
-        audit: audit_stage,
-        lint: lint_stage,
-        test: test_stage,
-        ci_profile: ci_profile_stage,
-    };
+    let (output, overall_exit) = ReviewService::output_from_stages(
+        ReviewOutputInput {
+            component: component_label.clone(),
+            plan: quality_plan,
+            observation: observation_metadata,
+            scope,
+            changed_since: args.changed_since.clone(),
+            changed_file_count,
+            head_ref: review::git_ref(&source_path, "HEAD").unwrap_or_default(),
+            hints: top_hints,
+        },
+        ReviewStages {
+            audit: audit_stage,
+            lint: lint_stage,
+            test: test_stage,
+            ci_profile: ci_profile_stage,
+        },
+    );
 
     print_human_summary(&output);
 
@@ -564,9 +375,9 @@ pub fn run_markdown(args: ReviewArgs, global: &GlobalArgs) -> CmdResult<String> 
     let banners = args.banner.clone();
     let (output, exit_code) = run(args, global)?;
     let md = if banners.is_empty() {
-        render::render_pr_comment(&output)
+        review::render::render_pr_comment(&output)
     } else {
-        render::render_pr_comment_with_banners(&output, &banners)
+        review::render::render_pr_comment_with_banners(&output, &banners)
     };
     Ok((md, exit_code))
 }
@@ -616,19 +427,6 @@ pub fn write_artifact_to_file(
     true
 }
 
-fn stage_skipped<T: Serialize>(stage: &str, reason: &str) -> ReviewStage<T> {
-    ReviewStage {
-        stage: stage.to_string(),
-        ran: false,
-        passed: true,
-        exit_code: 0,
-        finding_count: 0,
-        hint: format!("Run individually: homeboy {}", stage),
-        skipped_reason: Some(reason.to_string()),
-        output: None,
-    }
-}
-
 fn scope_flag_suffix(args: &ReviewArgs, include_changed_only: bool) -> String {
     if let Some(ref r) = args.changed_since {
         format!(" --changed-since={}", r)
@@ -637,93 +435,6 @@ fn scope_flag_suffix(args: &ReviewArgs, include_changed_only: bool) -> String {
     } else {
         String::new()
     }
-}
-
-fn build_artifact(
-    component: &str,
-    base_ref: &str,
-    head_ref: &str,
-    commands: Vec<ReviewArtifactCommand>,
-) -> ReviewArtifact {
-    let status = artifact_status(&commands).to_string();
-    ReviewArtifact {
-        schema: "homeboy/review/v1".to_string(),
-        component: component.to_string(),
-        status,
-        generated_at: generated_at_now(),
-        base_ref: base_ref.to_string(),
-        head_ref: head_ref.to_string(),
-        observation: None,
-        commands,
-    }
-}
-
-fn artifact_command<T: Serialize + ReviewArtifactFindings>(
-    stage: &ReviewStage<T>,
-) -> ReviewArtifactCommand {
-    ReviewArtifactCommand {
-        name: stage.stage.clone(),
-        status: if !stage.ran {
-            "skipped"
-        } else if stage.passed {
-            "passed"
-        } else {
-            "failed"
-        }
-        .to_string(),
-        exit_code: stage.exit_code,
-        summary: if !stage.ran {
-            stage
-                .skipped_reason
-                .clone()
-                .unwrap_or_else(|| "skipped".to_string())
-        } else {
-            format!(
-                "{} finding(s); {}",
-                stage.finding_count,
-                if stage.passed { "passed" } else { "failed" }
-            )
-        },
-        findings: stage
-            .output
-            .as_ref()
-            .map(ReviewArtifactFindings::review_artifact_findings)
-            .unwrap_or_default(),
-        artifacts: Vec::new(),
-    }
-}
-
-fn artifact_status(commands: &[ReviewArtifactCommand]) -> &'static str {
-    let ran = commands
-        .iter()
-        .filter(|command| command.status != "skipped")
-        .count();
-    if ran == 0 {
-        return "skipped";
-    }
-    if commands.iter().any(|command| command.status == "failed") {
-        return "failed";
-    }
-    if ran < commands.len() {
-        return "partial";
-    }
-    "passed"
-}
-
-fn generated_at_now() -> String {
-    Utc::now().to_rfc3339()
-}
-
-fn git_ref(path: &str, git_ref: &str) -> Option<String> {
-    let output = Command::new("git")
-        .args(["rev-parse", git_ref])
-        .current_dir(path)
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    Some(String::from_utf8_lossy(&output.stdout).trim().to_string())
 }
 
 fn build_audit_args(args: &ReviewArgs) -> audit::AuditArgs {
@@ -989,76 +700,6 @@ mod tests {
     }
 
     #[test]
-    fn stage_skipped_helper_marks_not_ran() {
-        let stage: ReviewStage<serde_json::Value> = stage_skipped("audit", "no files changed");
-        assert!(!stage.ran);
-        assert!(stage.passed);
-        assert_eq!(stage.exit_code, 0);
-        assert_eq!(stage.skipped_reason.as_deref(), Some("no files changed"));
-    }
-
-    #[test]
-    fn artifact_command_maps_stage_statuses() {
-        let skipped: ReviewStage<serde_json::Value> = stage_skipped("audit", "no files changed");
-        let skipped_command = artifact_command(&skipped);
-        assert_eq!(skipped_command.name, "audit");
-        assert_eq!(skipped_command.status, "skipped");
-        assert_eq!(skipped_command.exit_code, 0);
-        assert_eq!(skipped_command.summary, "no files changed");
-        assert!(skipped_command.findings.is_empty());
-        assert!(skipped_command.artifacts.is_empty());
-
-        let failed = ReviewStage {
-            stage: "lint".to_string(),
-            ran: true,
-            passed: false,
-            exit_code: 1,
-            finding_count: 3,
-            hint: "Deep dive: homeboy lint".to_string(),
-            skipped_reason: None,
-            output: Some(serde_json::json!({ "ok": false })),
-        };
-        let failed_command = artifact_command(&failed);
-        assert_eq!(failed_command.name, "lint");
-        assert_eq!(failed_command.status, "failed");
-        assert_eq!(failed_command.exit_code, 1);
-        assert_eq!(failed_command.summary, "3 finding(s); failed");
-    }
-
-    #[test]
-    fn artifact_status_covers_contract_values() {
-        let passed = ReviewArtifactCommand {
-            name: "lint".to_string(),
-            status: "passed".to_string(),
-            exit_code: 0,
-            summary: "0 finding(s); passed".to_string(),
-            findings: Vec::new(),
-            artifacts: Vec::new(),
-        };
-        let skipped = ReviewArtifactCommand {
-            name: "test".to_string(),
-            status: "skipped".to_string(),
-            exit_code: 0,
-            summary: "no files changed".to_string(),
-            findings: Vec::new(),
-            artifacts: Vec::new(),
-        };
-        let failed = ReviewArtifactCommand {
-            name: "audit".to_string(),
-            status: "failed".to_string(),
-            exit_code: 1,
-            summary: "1 finding(s); failed".to_string(),
-            findings: Vec::new(),
-            artifacts: Vec::new(),
-        };
-
-        assert_eq!(artifact_status(std::slice::from_ref(&skipped)), "skipped");
-        assert_eq!(artifact_status(std::slice::from_ref(&passed)), "passed");
-        assert_eq!(artifact_status(&[passed.clone(), skipped]), "partial");
-        assert_eq!(artifact_status(&[passed, failed]), "failed");
-    }
-
-    #[test]
     fn unsupported_review_plan_step_returns_consistent_error() {
         let args = review_args_fixture();
         let global = GlobalArgs {};
@@ -1072,93 +713,6 @@ mod tests {
         assert!(err
             .to_string()
             .contains("unsupported executable step 'review.unknown'"));
-    }
-
-    #[test]
-    fn execute_review_plan_steps_preserves_quality_order() {
-        let steps = vec![
-            PlanStep::ready("review.audit", "review.audit").build(),
-            PlanStep::ready("review.lint", "review.lint").build(),
-            PlanStep::ready("review.test", "review.test").build(),
-        ];
-        let mut observed = Vec::new();
-
-        let run = execute_review_plan_steps(&steps, |step| {
-            observed.push(step.id.clone());
-            Ok(Some(step.id.clone()))
-        })
-        .expect("review execution should run ready steps");
-
-        assert_eq!(observed, vec!["review.audit", "review.lint", "review.test"]);
-        assert_eq!(run.results, observed);
-        assert!(!run.stopped);
-    }
-
-    #[test]
-    fn execute_review_plan_steps_skips_disabled_and_skipped_steps() {
-        let steps = vec![
-            PlanStep::ready("review.audit", "review.audit").build(),
-            PlanStep::disabled_with_reason("review.lint", "review.lint", "disabled").build(),
-            PlanStep::builder(
-                "review.test",
-                "review.test",
-                homeboy::core::plan::PlanStepStatus::Skipped,
-            )
-            .build(),
-        ];
-
-        let run = execute_review_plan_steps(&steps, |step| Ok(Some(step.id.clone())))
-            .expect("review execution should ignore non-executable steps");
-
-        assert_eq!(run.results, vec!["review.audit"]);
-        assert!(!run.stopped);
-    }
-
-    #[test]
-    fn execute_review_plan_steps_does_not_treat_failures_as_show_stoppers() {
-        let steps = vec![
-            PlanStep::ready("review.audit", "review.audit").build(),
-            PlanStep::ready("review.lint", "review.lint").build(),
-            PlanStep::ready("review.test", "review.test").build(),
-        ];
-
-        let run = execute_review_plan_steps(&steps, |step| {
-            let exit_code = if step.id == "review.audit" { 2 } else { 0 };
-            Ok(Some((step.id.clone(), exit_code)))
-        })
-        .expect("review execution should continue after stage failures");
-
-        assert_eq!(
-            run.results,
-            vec![
-                ("review.audit".to_string(), 2),
-                ("review.lint".to_string(), 0),
-                ("review.test".to_string(), 0),
-            ]
-        );
-        assert!(!run.stopped);
-    }
-
-    #[test]
-    fn build_artifact_uses_review_schema_and_refs() {
-        let command = ReviewArtifactCommand {
-            name: "lint".to_string(),
-            status: "passed".to_string(),
-            exit_code: 0,
-            summary: "0 finding(s); passed".to_string(),
-            findings: Vec::new(),
-            artifacts: Vec::new(),
-        };
-
-        let artifact = build_artifact("homeboy", "origin/main", "abc123", vec![command]);
-
-        assert_eq!(artifact.schema, "homeboy/review/v1");
-        assert_eq!(artifact.component, "homeboy");
-        assert_eq!(artifact.status, "passed");
-        assert_eq!(artifact.base_ref, "origin/main");
-        assert_eq!(artifact.head_ref, "abc123");
-        assert_eq!(artifact.commands.len(), 1);
-        assert!(artifact.generated_at.contains('T'));
     }
 
     #[test]
